+++ enclone_visual/src/gui.rs
@@ -1,9 +1,7 @@
 // Copyright (c) 2021 10X Genomics, Inc. All rights reserved.
 
-use crate::convert_svg_to_png::*;
-use crate::copy_image_to_clipboard::*;
-use crate::svg_to_geometry::*;
 use crate::*;
+use crate::copy_image_to_clipboard::*;
 use gui_structures::*;
 use gui_structures::ComputeState::*;
 use iced::svg::Handle;
@@ -31,36 +29,6 @@ const DEJAVU_BOLD: Font = Font::External {
 pub static COUNT: AtomicUsize = AtomicUsize::new(0);
 
 // ▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓
-
-impl EncloneVisual {
-    pub fn post_svg(&mut self, svg: &str) {
-        self.png_value = convert_svg_to_png(&svg.as_bytes());
-        let geometry = svg_to_geometry(&svg, false);
-        if geometry.is_some() {
-            let mut ok = true;
-            for i in 0..geometry.as_ref().unwrap().len() {
-                match &geometry.as_ref().unwrap()[i] {
-                    crate::geometry::Geometry::Text(ttt) => {
-                        if ttt.rotate != [0.0; 3] {
-                            ok = false;
-                        }
-                    }
-                    _ => {}
-                }
-            }
-            if ok {
-                self.canvas_view.state.geometry_value = geometry;
-            } else {
-                self.canvas_view.state.geometry_value = None;
-            }
-        } else {
-            if VERBOSE.load(SeqCst) {
-                println!("translation from svg to geometries failed");
-            }
-            self.canvas_view.state.geometry_value = None;
-        }
-    }
-}
 
 impl Application for EncloneVisual {
     type Executor = iced::executor::Default;
